+++ eras/dijkstra/impl/CHANGELOG.md
@@ -120,6 +120,8 @@
 
 ### cddl
 
+* Remove re-exported `genByteString`, `distinct`, `genHash28`, `majorProtocolVersionRule`, `ipRule` and `ipValidator`
+* Remove `dijkstraProtocolVersionRule`
 * Add `transaction_mempool` rule
 * Add `peras_certificate`, `block_body`
 * Extend `constr` CDDL rule to include tags 1280–1400 for Plutus `Data` constructor indexes