+++ eras/mary/impl/CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### `cddl`
 
+* Remove re-exported `genByteString`, `distinct`, `genHash28`, `majorProtocolVersionRule`, `ipRule` and `ipValidator`
 * Change `ipv4` and `ipv6` to use exact byte sizes (4 and 16 respectively), no longer allowing leftover bytes
 
 ### `testlib`